// dashboard/src/api/models/couponsListStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */

export type CouponsListStatus =
  (typeof CouponsListStatus)[keyof typeof CouponsListStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CouponsListStatus = {
  active: "active",
  archived: "archived",
} as const;