// dashboard/src/api/models/pricesListStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */

export type PricesListStatus =
  (typeof PricesListStatus)[keyof typeof PricesListStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PricesListStatus = {
  active: "active",
  archived: "archived",
} as const;