// dashboard/src/api/models/quoteLine.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */
import type { QuoteLineDiscount } from "./quoteLineDiscount";
import type { QuoteLineTax } from "./quoteLineTax";

export interface QuoteLine {
  id: string;
  description: string;
  quantity: number;
  /** @pattern ^-?\d{0,17}(?:\.\d{0,2})?$ */
  unit_amount: string;
  /** @nullable */
  price_id: string | null;
  /** @nullable */
  product_id: string | null;
  /** @pattern ^-?\d{0,17}(?:\.\d{0,2})?$ */
  readonly amount: string;
  /** @pattern ^-?\d{0,17}(?:\.\d{0,2})?$ */
  readonly total_discount_amount: string;
  /** @pattern ^-?\d{0,17}(?:\.\d{0,2})?$ */
  readonly total_amount_excluding_tax: string;
  /** @pattern ^-?\d{0,17}(?:\.\d{0,2})?$ */
  readonly total_tax_amount: string;
  /** @pattern ^-?\d{0,3}(?:\.\d{0,2})?$ */
  readonly total_tax_rate: string;
  /** @pattern ^-?\d{0,17}(?:\.\d{0,2})?$ */
  readonly total_amount: string;
  discounts: QuoteLineDiscount[];
  taxes: QuoteLineTax[];
}